Avoid unnecessary error output for cookbooks without tests.
The remote_directory and cookbook_file resources uses ignore_failure
to account for cookbooks that do not have tests or support files.
However, with the addition of error inspectors, the errors these
resources raise produce a large amount of output that obfuscates the
test results.  The following link has an example of the problematic
output:

https://gist.github.com/4191403

This commit tests for the exists of the tests and support and run-time
and produces more consise output:

    WARN: No tests found for test-kitchen::default!
    DEBUG: No support files found for test-kitchen::default.

Note that this depends on internal implementation details of the
Chef::CookbookVersion class which may increase the maintenance burden
of this cookbook.